requirements.txt
opencv-python==4.11.0.86
qrcode==8.2
pillow==11.2.1
pytest==8.3.5
fpdf2==2.8.3
